Switch to `HasResult` and refactor name
