+++ patches/minecraft/net/minecraft/inventory/Slot.java.patch
@@ -9,18 +9,17 @@
      }
  
      public ItemStack func_75209_a(int p_75209_1_)
-@@ -105,7 +105,9 @@
+@@ -105,7 +105,8 @@
  
      public boolean func_82869_a(EntityPlayer p_82869_1_)
      {
 -        return true;
 +        ItemStack itemStack = this.func_75211_c();
-+        net.minecraft.util.EnumActionResult cancelResult = net.minecraftforge.common.ForgeHooks.onItemTake(p_82869_1_, itemStack);
-+        return (cancelResult == null);
++        return net.minecraftforge.event.ForgeEventFactory.onCanTakeItem(p_82869_1_, itemStack);
      }
  
      @SideOnly(Side.CLIENT)
-@@ -113,4 +115,74 @@
+@@ -113,4 +114,74 @@
      {
          return true;
      }
