+++ src/main/java/net/minecraftforge/items/SlotItemHandler.java
@@ -126,8 +126,8 @@ public class SlotItemHandler extends Slot
     @Override
     public boolean canTakeStack(EntityPlayer playerIn)
     {
-        net.minecraft.util.EnumActionResult cancelResult = net.minecraftforge.common.ForgeHooks.onItemTake(playerIn, this.getStack());
-        return cancelResult == null && !this.getItemHandler().extractItem(index, 1, true).isEmpty();
+        boolean result = net.minecraftforge.event.ForgeEventFactory.onCanTakeItem(playerIn, this.getStack());
+        return result && !this.getItemHandler().extractItem(index, 1, true).isEmpty();
     }
 
     @Override
